Move menu functions to roles modules
