--- gerente.py
@@ -1,70 +1,0 @@
-from __future__ import annotations
-
-from conexion.conexion import ConexionBD
-
-
-def menu_gerente(conexion: ConexionBD, correo: str, es_admin: bool = False) -> None:
-    while True:
-        print("\n--- Menú Gerente ---")
-        print("1. Gestionar empleados")
-        print("2. Ver reportes")
-        print("3. Gestionar inventario de vehículos")
-        print("4. Asignar tareas o roles")
-        print("5. Ver historial de operaciones")
-        print("6. Estadísticas de clientes frecuentes")
-        print("7. Generar informes")
-        print("8. Gestionar notificaciones")
-        print("0. Salir")
-        op = input("Seleccione una opción: ").strip()
-        if op == "1":
-            gestionar_empleados(conexion)
-        elif op == "2":
-            ver_reportes(conexion)
-        elif op == "3":
-            gestionar_inventario(conexion)
-        elif op == "4":
-            asignar_tareas(conexion)
-        elif op == "5":
-            historial_operaciones(conexion)
-        elif op == "6":
-            estadisticas_clientes(conexion)
-        elif op == "7":
-            generar_informes(conexion)
-        elif op == "8":
-            gestionar_notificaciones(conexion)
-        elif op == "0":
-            break
-        else:
-            print("Opción inválida")
-
-
-def gestionar_empleados(conexion: ConexionBD) -> None:
-    print("Función de gestión de empleados no implementada")
-
-
-def ver_reportes(conexion: ConexionBD) -> None:
-    print("Función de reportes no implementada")
-
-
-def gestionar_inventario(conexion: ConexionBD) -> None:
-    print("Función de inventario no implementada")
-
-
-def asignar_tareas(conexion: ConexionBD) -> None:
-    print("Función de asignación de tareas no implementada")
-
-
-def historial_operaciones(conexion: ConexionBD) -> None:
-    print("Función de historial de operaciones no implementada")
-
-
-def estadisticas_clientes(conexion: ConexionBD) -> None:
-    print("Función de estadísticas no implementada")
-
-
-def generar_informes(conexion: ConexionBD) -> None:
-    print("Función de generación de informes no implementada")
-
-
-def gestionar_notificaciones(conexion: ConexionBD) -> None:
-    print("Función de notificaciones no implementada")
